Rename EntityType model to ResourceType
